fixed the creation of the identity and zero matrix

+++ headers/Matrix3D.h
@@ -9,7 +9,7 @@
 class Matrix3D
 {
 public:
-    Matrix3D(std::vector<std::vector<int>> matrix) :
+    Matrix3D(std::vector<std::vector<float>> matrix) :
         matrix(matrix) {}
 
     Matrix3D create_zero_matrix();
